requirements.txt
torch==2.0.0+cu118  
torchvision==0.15.0+cu118  
numpy
matplotlib
Pillow
kaggle
